requirements.txt
beautifulsoup4==4.12.3
deepl==1.18.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-experimental==0.0.55
langchain-openai==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.31
lxml==5.2.1
python-dotenv==1.0.1
pytube==15.0.0
streamlit==1.33.0
st-pages==0.4.5
streamlit-tags==1.2.8
spacy==3.7.4
youtube_transcript_api==0.6.2
